Re-enable TMVar
